+++ src/common/common.h
@@ -43,7 +43,7 @@ struct flow_metadata
  */
 
 /* read exactly 'count' bytes from a socket 'fd' */
-unsigned int read_exact(int fd, char *buf, size_t count, size_t max_per_read, bool dummy_buf);
+unsigned long long read_exact(int fd, char *buf, size_t count, size_t max_per_read, bool dummy_buf);
 
 /* write exactly 'count' bytes into a socket 'fd' */
 unsigned long long write_exact(int fd, char *buf, size_t count, size_t max_per_write,
